fix(Mesh): move some files in private headers
