+++ include/geode/mesh/core/private/solid_mesh_view_impl.h
@@ -26,8 +26,8 @@
 #include <geode/basic/attribute_manager.h>
 
 #include <geode/mesh/builder/solid_mesh_builder.h>
-#include <geode/mesh/core/detail/points_view_impl.h>
 #include <geode/mesh/core/detail/vertex_cycle.h>
+#include <geode/mesh/core/private/points_view_impl.h>
 #include <geode/mesh/core/solid_mesh.h>
 
 namespace geode
