+++ src/geode/mesh/core/geode_edged_curve.cpp
@@ -32,8 +32,8 @@
 
 #include <geode/geometry/point.h>
 
-#include <geode/mesh/core/detail/edges_impl.h>
-#include <geode/mesh/core/detail/points_impl.h>
+#include <geode/mesh/core/private/edges_impl.h>
+#include <geode/mesh/core/private/points_impl.h>
 
 namespace geode
 {
