+++ src/geode/mesh/core/geode_point_set.cpp
@@ -30,7 +30,7 @@
 
 #include <geode/geometry/point.h>
 
-#include <geode/mesh/core/detail/points_impl.h>
+#include <geode/mesh/core/private/points_impl.h>
 
 namespace geode
 {
