+++ src/geode/mesh/core/geode_polygonal_surface.cpp
@@ -32,7 +32,7 @@
 
 #include <geode/geometry/point.h>
 
-#include <geode/mesh/core/detail/points_impl.h>
+#include <geode/mesh/core/private/points_impl.h>
 
 namespace geode
 {
