+++ src/geode/mesh/core/geode_tetrahedral_solid.cpp
@@ -35,7 +35,7 @@
 #include <geode/geometry/point.h>
 
 #include <geode/mesh/core/detail/geode_elements.h>
-#include <geode/mesh/core/detail/points_impl.h>
+#include <geode/mesh/core/private/points_impl.h>
 
 namespace geode
 {
