+++ src/geode/mesh/core/solid_edges.cpp
@@ -37,9 +37,9 @@
 
 #include <geode/mesh/builder/solid_mesh_builder.h>
 #include <geode/mesh/core/bitsery_archive.h>
-#include <geode/mesh/core/detail/facet_edges_impl.h>
 #include <geode/mesh/core/mesh_factory.h>
 #include <geode/mesh/core/polyhedral_solid.h>
+#include <geode/mesh/core/private/facet_edges_impl.h>
 
 namespace
 {
